Refactor to simplify struct and get edges right
The state structure is now a simple namedtuple, to discourage users from
fiddling its content through methods. Its previous logic is merged into
the resolution class.

The pinning process is revised to ensure there are no stale edges in the
graph when we re-pin a different candidate to the same name.