Re-implement backtracking state management
The whole backtrack() function is picked apart and put together again. I
think this finally makes 100% sense to me, my previous implementation
was such a flux.

Every round of resolution is now a process of

1. Inheriting criteria from the prvious state
2. Create a new pin

When backtracking, the state to backtrack into not only inherits
criteria from the "parent" state, but also receives additional
incompatibility information from the "child" state that discovered them.
So when we backtrack into it, the state's criteria should be modified
in-place so subsequence backtracking into it would also know about those
incompatibilities.

Also (this is the part I got wrong), since the state being backtracked
into does not produce a valid pin after the backtracking process (its
previous pin is now considered an incompatibility), we SHOULD NOT push a
new state; the next round should work on producing a pin on that state.