Implement backjumping and skip unrelated states
The current backtracking logic assumes that the last state in the stack
is the state that caused the incompatibility. This is not always the
case. For example, given three requirements A, B and C, with
dependencies A1, B1 and C1, where A1 and B1 are incompatible. The
requirements are processed one after the other, so the last state is
related to C, while the incompatibility is caused by B.

The current behaviour causes significant slowdowns in case there are
many candidates for B and C, as all their combination are evaluated
before a compatible version of B can be found.

The new backjumping behaviour discards states iteratively "backward"
until any of the candidates in the state depends on any of the packages
that caused the incompatibility in the current state. In our example,
this causes the state related to C to be dropped without evaluating any
of its candidates, until a compatible candidate of B is found.

If the backjumping logic is not able to find a compatible candidate,
it means that there is no solution for the requirements and a
ResolutionImpossible exception is raised.

Co-authored-by: Frost Ming <mianghong@gmail.com>
Co-authored-by: Pradyun Gedam <pradyunsg@gmail.com>
Signed-off-by: Bennati, Stefano <stefano.bennati@here.com>