Change `Mapping` to `dict` in resolution `Result` type annotations
`typing.Mapping` is a broad concept. The actual code just assigns `dict`s so `Result` should just use that.

I've been adding annotations to the calling side (`ansible-galaxy`) that uses `resolvelib` and hit this as I was trying to show that a function that returns `Resolver(...).resolve(...).mapping` produces a `dict`.